move start_lightning to coins_activation to use LightningRpcTaskHandle to update the enabling progress
