+++ mm2src/coins/lightning/ln_errors.rs
@@ -4,6 +4,7 @@ use db_common::sqlite::rusqlite::Error as SqlError;
 use derive_more::Display;
 use http::StatusCode;
 use mm2_err_handle::prelude::*;
+use rpc_task::RpcTaskError;
 use std::num::TryFromIntError;
 
 pub type EnableLightningResult<T> = Result<T, MmError<EnableLightningError>>;
@@ -30,6 +31,8 @@ pub enum EnableLightningError {
     RpcError(String),
     #[display(fmt = "DB error {}", _0)]
     DbError(String),
+    #[display(fmt = "Rpc task error: {}", _0)]
+    RpcTaskError(String),
     ConnectToNodeError(String),
 }
 
@@ -44,7 +47,8 @@ impl HttpStatusCode for EnableLightningError {
             | EnableLightningError::IOError(_)
             | EnableLightningError::ConnectToNodeError(_)
             | EnableLightningError::InvalidConfiguration(_)
-            | EnableLightningError::DbError(_) => StatusCode::INTERNAL_SERVER_ERROR,
+            | EnableLightningError::DbError(_)
+            | EnableLightningError::RpcTaskError(_) => StatusCode::INTERNAL_SERVER_ERROR,
         }
     }
 }
@@ -59,6 +63,10 @@ impl From<SqlError> for EnableLightningError {
 
 impl From<UtxoRpcError> for EnableLightningError {
     fn from(e: UtxoRpcError) -> Self { EnableLightningError::RpcError(e.to_string()) }
+}
+
+impl From<RpcTaskError> for EnableLightningError {
+    fn from(e: RpcTaskError) -> Self { EnableLightningError::RpcTaskError(e.to_string()) }
 }
 
 #[derive(Display, PartialEq)]
