+++ mm2src/coins_activation/Cargo.toml
@@ -18,6 +18,7 @@ mm2_core = { path = "../mm2_core" }
 mm2_err_handle = { path = "../mm2_err_handle" }
 mm2_metrics = { path = "../mm2_metrics" }
 mm2_number = { path = "../mm2_number" }
+parking_lot = { version = "0.12.0", features = ["nightly"] }
 rpc = { path = "../mm2_bitcoin/rpc" }
 rpc_task = { path = "../rpc_task" }
 ser_error = { path = "../derives/ser_error" }
@@ -25,3 +26,8 @@ ser_error_derive = { path = "../derives/ser_error_derive" }
 serde = "1.0"
 serde_derive = "1.0"
 serde_json = { version = "1.0", features = ["preserve_order", "raw_value"] }
+
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+lightning = "0.0.110"
+lightning-background-processor = "0.0.110"
+lightning-invoice = { version = "0.18.0", features = ["serde"] }
